Do not reuse chrome instances between runs
Previously, it was set to reuse them to save time. Sometime between
Christmas and now, the chrome people changed something in chrome (or
chrome devtools protocol) that makes it so that lighthouse can no longer
reuse chrome instances. When I got back it just stopped working (I
didn't change the lighthouse version but chrome probably auto-updated).

[I opened an issue on lighthouse](https://github.com/GoogleChrome/lighthouse/issues/14736),
but there does not seem to be an easy resolution (other than that their
@next version works). I am reluctant to make lighthouse-parade dependent
on the lighthouse prerelease version.

So, for now, each chrome instance is used once and then thrown away.